Introduces getColumn and removeColumn methods to TableBuilderAbstract to make it easier to change format of a table.

+++ cloudsim-plus/src/main/java/org/cloudsimplus/builders/tables/CloudletsTableBuilder.java
@@ -41,7 +41,7 @@ import java.util.List;
  * @since CloudSim Plus 1.0
  */
 public class CloudletsTableBuilder extends TableBuilderAbstract<Cloudlet> {
-    private static final String TIME_FORMAT = "%d";
+    private static final String TIME_FORMAT = "%.0f";
     private static final String SECONDS = "Seconds";
     private static final String CPU_CORES = "CPU cores";
 
@@ -81,12 +81,12 @@ public class CloudletsTableBuilder extends TableBuilderAbstract<Cloudlet> {
         addColumnDataFunction(getTable().addColumn("CloudletPEs", CPU_CORES), Cloudlet::getNumberOfPes);
 
         TableColumn col = getTable().addColumn("StartTime", SECONDS).setFormat(TIME_FORMAT);
-        addColumnDataFunction(col, cloudlet -> (long)cloudlet.getExecStartTime());
+        addColumnDataFunction(col, cloudlet -> cloudlet.getExecStartTime());
 
         col = getTable().addColumn("FinishTime", SECONDS).setFormat(TIME_FORMAT);
-        addColumnDataFunction(col, cloudlet -> (long)cloudlet.getFinishTime());
+        addColumnDataFunction(col, cloudlet -> cloudlet.getFinishTime());
 
         col = getTable().addColumn("ExecTime", SECONDS).setFormat(TIME_FORMAT);
-        addColumnDataFunction(col, cloudlet -> (long)Math.ceil(cloudlet.getActualCpuTime()));
+        addColumnDataFunction(col, cloudlet -> Math.ceil(cloudlet.getActualCpuTime()));
     }
 }
